Move node-tree update manager to the managers package
