+++ services/repository-managers/src/main/java/org/sagebionetworks/repo/manager/dynamo/NodeTreeUpdateManager.java
@@ -1,4 +1,4 @@
-package org.sagebionetworks.dynamo.manager;
+package org.sagebionetworks.repo.manager.dynamo;
 
 import java.util.Date;
