+++ services/repository-managers/src/main/java/org/sagebionetworks/repo/manager/dynamo/NodeTreeUpdateManagerImpl.java
@@ -1,4 +1,4 @@
-package org.sagebionetworks.dynamo.manager;
+package org.sagebionetworks.repo.manager.dynamo;
 
 import java.util.ArrayList;
 import java.util.Date;
