+++ services/repository-managers/src/main/resources/managers-spb.xml
@@ -217,6 +217,11 @@
 	<bean id="nodeTreeQueryManager" class="org.sagebionetworks.repo.manager.dynamo.NodeTreeQueryManagerImpl"
 		scope="singleton" />
 
+	<bean id="nodeTreeUpdateManager" class="org.sagebionetworks.repo.manager.dynamo.NodeTreeUpdateManagerImpl">
+		<constructor-arg index="0" ref="nodeTreeUpdateDao" />
+		<constructor-arg index="1" ref="nodeDao" />
+	</bean>
+
 	<bean id="storageQuotaManager" class="org.sagebionetworks.repo.manager.StorageQuotaManagerImpl"
 		scope="singleton" />
